# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey API",
        "description": "REST API сервиса опросов и голосований",
        "contact": {
            "name": "API Support",
            "email": "support@survey.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/surveys/{survey}/analytics": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Получить аналитику по опросу (статистика ответов)",
                "operationId": "6cf470efb6418adf1b2e24ecafbe90e0",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Аналитика опроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "survey_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Опрос о языках"
                                        },
                                        "total_respondents": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "questions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "question_id": {
                                                        "type": "integer"
                                                    },
                                                    "question": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "single",
                                                            "multiple",
                                                            "text"
                                                        ]
                                                    },
                                                    "answers": {
                                                        "description": "Для text вопросов",
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "options": {
                                                        "description": "Для choice вопросов",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "option_id": {
                                                                    "type": "integer"
                                                                },
                                                                "option": {
                                                                    "type": "string"
                                                                },
                                                                "count": {
                                                                    "type": "integer"
                                                                },
                                                                "percent": {
                                                                    "type": "number",
                                                                    "example": 23.8
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен - не автор опроса"
                    },
                    "404": {
                        "description": "Опрос не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys/{survey}/export": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Экспортировать полные данные опроса",
                "operationId": "9de88cfc6e9c0477bccb7e84b0c76809",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Полная структура опроса с ответами пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "draft",
                                                "published",
                                                "closed"
                                            ]
                                        },
                                        "questions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        },
                                        "responses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Response"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Опрос не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Пользователь",
                "description": "Модель пользователя системы",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Дима"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "dima@test.com"
                    },
                    "role": {
                        "description": "Роль пользователя",
                        "type": "string",
                        "enum": [
                            "author",
                            "respondent"
                        ],
                        "example": "author"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Question": {
                "title": "Вопрос",
                "description": "Вопрос в опросе",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "survey_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Какой ваш любимый язык?"
                    },
                    "type": {
                        "description": "Тип вопроса: single (одиночный выбор), multiple (множественный выбор), text (текстовый)",
                        "type": "string",
                        "enum": [
                            "single",
                            "multiple",
                            "text"
                        ],
                        "example": "single"
                    },
                    "order": {
                        "description": "Порядок вопроса",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "options": {
                        "description": "Варианты ответов (только для single/multiple)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        }
                    }
                },
                "type": "object"
            },
            "Option": {
                "title": "Вариант ответа",
                "description": "Вариант ответа на вопрос",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "question_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "PHP"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Response": {
                "title": "Прохождение опроса пользователем",
                "description": "Запись о прохождении пользователем опроса",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "survey_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "answers": {
                        "description": "Ответы пользователя",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Answer": {
                "title": "Ответ на вопрос",
                "description": "Ответ пользователя на конкретный вопрос",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "response_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "question_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "option_id": {
                        "description": "ID выбранного варианта (null для текстовых ответов)",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "text_value": {
                        "description": "Текстовый ответ (null для выбора вариантов)",
                        "type": "string",
                        "example": "Другой язык",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "question": {
                        "$ref": "#/components/schemas/Question"
                    },
                    "option": {
                        "$ref": "#/components/schemas/Option"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT токен в формате: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}